Switch from Travis CI to GitHub actions (#52)
Closes #40

The build is passing on Node 8, 10, 12, 14 under both Linux and Windows. The badge has been updated, Travis CI and AppVeyor removed.

The only issue I couldn't solve is the scenario testing debug(). It fails on Windows, no output seems to be caught on stdout/stderr. I've added a @skip-windows tag to ignore it on CI when running under windows.